--- backend/src/main/java/kr/sosu/dailyline/DailylineApplication.java
@@ -1,13 +1,0 @@
-package kr.sosu.dailyline;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DailylineApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DailylineApplication.class, args);
-	}
-
-}
