--- backend/src/main/java/kr/sosu/dailyline/domain/user/controller/HomeController.java
@@ -1,25 +1,0 @@
-package kr.sosu.dailyline.domain.user.controller;
-
-import kr.sosu.dailyline.global.auth.LoginUser;
-import kr.sosu.dailyline.global.auth.entity.SessionUser;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-
-@Slf4j
-@RestController
-public class HomeController {
-
-    @GetMapping("/")
-    public void home(@LoginUser SessionUser sessionUser) {
-        // 리팩토링 예정
-        if(sessionUser != null) {
-            System.out.println("로그인");
-            System.out.println(sessionUser.getUserEmail());
-            System.out.println(sessionUser.getUserName());
-        } else {
-            System.out.println("로그아웃");
-        }
-    }
-}
