--- backend/src/main/java/kr/sosu/dailyline/domain/user/db/entity/Role.java
@@ -1,13 +1,0 @@
-package kr.sosu.dailyline.domain.user.db.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-    USER("ROLE_USER", "회원");
-
-    private final String key;
-    private final String title;
-}
