--- backend/src/main/java/kr/sosu/dailyline/domain/user/db/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package kr.sosu.dailyline.domain.user.db.repository;
-
-import kr.sosu.dailyline.domain.user.db.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUserEmail(String userEmail);
-}
