--- backend/src/main/java/kr/sosu/dailyline/domain/user/service/UserServiceImpl.java
@@ -1,7 +1,0 @@
-package kr.sosu.dailyline.domain.user.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserServiceImpl implements UserService {
-}
