--- backend/src/main/java/kr/sosu/dailyline/global/auth/LoginUser.java
@@ -1,11 +1,0 @@
-package kr.sosu.dailyline.global.auth;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface LoginUser {
-}
