--- backend/src/main/java/kr/sosu/dailyline/global/config/SecurityConfig.java
@@ -1,34 +1,0 @@
-package kr.sosu.dailyline.global.config;
-
-import kr.sosu.dailyline.domain.user.db.entity.Role;
-import kr.sosu.dailyline.global.auth.CustomOAuthUserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
-@RequiredArgsConstructor
-@EnableWebSecurity
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    private final CustomOAuthUserService customOAuthUserService;
-
-    @Override
-    protected void configure(HttpSecurity httpSecurity) throws Exception {
-        httpSecurity
-                .csrf().disable()
-                .headers().frameOptions().disable()
-                .and()
-                .authorizeRequests()
-                .antMatchers("/", "/css/**", "/images/**", "/js/**", "/swagger-ui/**", "/v3/api-docs/**").permitAll()
-                .antMatchers("/api/users/**").hasRole(Role.USER.name())
-                .anyRequest().authenticated()
-                .and()
-                    .logout()
-                        .logoutSuccessUrl("/")
-                .and()
-                    .oauth2Login()
-                        .userInfoEndpoint()
-                            .userService(customOAuthUserService);
-    }
-}
