--- backend/src/main/java/kr/sosu/dailyline/global/config/SwaggerConfig.java
@@ -1,26 +1,0 @@
-package kr.sosu.dailyline.global.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springdoc.core.GroupedOpenApi;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public GroupedOpenApi publicApi() {
-        return GroupedOpenApi.builder()
-                .group("daily-line")
-                .pathsToMatch("/api/**")
-                .build();
-    }
-    @Bean
-    public OpenAPI springShopOpenAPI() {
-        return new OpenAPI()
-                .info(new Info().title("Daily Line API")
-                        .description("Daily Line API 명세서입니다.")
-                        .version("v0.0.1"));
-    }
-}
