+++ backend/src/main/java/kr/sosu/project/DailylineApplication.java
@@ -1,0 +1,13 @@
+package kr.sosu.project;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DailylineApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DailylineApplication.class, args);
+	}
+
+}
