+++ backend/src/main/java/kr/sosu/project/dailyline/challenge/repository/ChallengeRepository.java
@@ -1,0 +1,9 @@
+package kr.sosu.project.dailyline.challenge.repository;
+
+import kr.sosu.project.dailyline.challenge.domain.Challenge;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
+}
