+++ backend/src/main/java/kr/sosu/project/dailyline/user/domain/Role.java
@@ -1,0 +1,13 @@
+package kr.sosu.project.dailyline.user.domain;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Role {
+    USER("ROLE_USER", "회원");
+
+    private final String key;
+    private final String title;
+}
