+++ backend/src/main/java/kr/sosu/project/dailyline/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package kr.sosu.project.dailyline.user.repository;
+
+import kr.sosu.project.dailyline.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUserEmail(String userEmail);
+}
