+++ backend/src/main/java/kr/sosu/project/dailyline/user/service/UserServiceImpl.java
@@ -1,0 +1,7 @@
+package kr.sosu.project.dailyline.user.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserServiceImpl implements UserService {
+}
