+++ backend/src/main/java/kr/sosu/project/global/auth/LoginUser.java
@@ -1,0 +1,11 @@
+package kr.sosu.project.global.auth;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LoginUser {
+}
